rpgweb/pip_requirements.txt

Django>=1.3
django-debug-toolbar>=0.8.5
ZODB3>=3.10.3
django_zodb
PyYAML>=3.10
decorator>=3.3.1
odict>=1.4.1
pyparsing>=1.5.6
#Jinja2==2.5.5
#Coffin==0.3.4



